test: migrate zero-billing-portal + zero-billing-invoices to bdd
